Enable fastify opts in plugins
