+++ packages/web/config/logger-options.js
@@ -1,0 +1,32 @@
+/**
+ * @import { FastifyServerOptions } from 'fastify'
+ */
+
+const PinoLevelToSeverityLookup = /** @type {const} */ ({
+  trace: 'DEBUG',
+  debug: 'DEBUG',
+  info: 'INFO',
+  warn: 'WARNING',
+  error: 'ERROR',
+  fatal: 'CRITICAL',
+})
+
+export const loggerOptions = /** @type{const} @satisfies {FastifyServerOptions['logger']} */ ({
+  mixin () {
+    return {
+      service: 'bc-web',
+    }
+  },
+  messageKey: 'message',
+  formatters: {
+    level (/** @type{string} */label, /** @type{number} */number) {
+      return {
+        level: PinoLevelToSeverityLookup[
+          /** @type {keyof typeof PinoLevelToSeverityLookup} */
+          (label)
+        ] || PinoLevelToSeverityLookup.info,
+        levelN: number,
+      }
+    },
+  },
+})
