+++ packages/web/fly.toml
@@ -36,7 +36,6 @@ swap_size_mb = 1024
   PORT = "8080"
   PQUEUE_CONCURRENCY = "2"
   TRANSPORT = "https"
-  METRICS = "1"
   OTEL_SERVICE_NAME = "breadcrum-web"
   OTEL_SERVICE_VERSION = "1.0.0"
   OTEL_RESOURCE_ATTRIBUTES = "deployment.environment=production"
