+++ packages/web/plugins/pkg-decorators.d.ts
@@ -1,0 +1,8 @@
+import 'fastify'
+import type { PackageJson } from 'type-fest'
+
+declare module 'fastify' {
+  interface FastifyInstance {
+    pkg: PackageJson
+  }
+}
